+++ packages/create-gatsby/src/init-starter.ts
@@ -7,7 +7,7 @@ import { spin } from "tiny-spin"
 import { getConfigStore } from "./get-config-store"
 type PackageManager = "yarn" | "npm"
 import c from "ansi-colors"
-import { clearLine, kebabify } from "./utils"
+import { clearLine, makeNpmSafe } from "./utils"
 
 const packageManagerConfigKey = `cli.packageManager`
 
@@ -85,8 +85,9 @@ const setNameInPackage = async (
 ): Promise<void> => {
   const packageJsonPath = path.join(sitePath, `package.json`)
   const packageJson = await fs.readJSON(packageJsonPath)
-  packageJson.name = kebabify(name)
+  packageJson.name = makeNpmSafe(name)
   packageJson.description = name
+  delete packageJson.license
   try {
     const result = await execa(`git`, [`config`, `user.name`])
     if (result.failed) {
@@ -178,6 +179,8 @@ const clone = async (
 
   stop()
   await fs.remove(path.join(rootPath, `.git`))
+
+  await fs.remove(path.join(rootPath, `LICENSE`))
 }
 
 export async function gitSetup(rootPath: string): Promise<void> {