+++ packages/gatsby-plugin-sharp/src/image-data.ts
@@ -9,7 +9,7 @@ import { createTransformObject } from "./plugin-options"
 
 const DEFAULT_BLURRED_IMAGE_WIDTH = 20
 
-type ImageFormat = "jpg" | "png" | "webp" | "" | "auto"
+type ImageFormat = "jpg" | "png" | "webp" | "avif" | "" | "auto"
 export interface ISharpGatsbyImageArgs {
   layout?: "fixed" | "fluid" | "constrained"
   formats?: Array<ImageFormat>
@@ -28,6 +28,7 @@ export interface ISharpGatsbyImageArgs {
   jpgOptions: Record<string, unknown>
   pngOptions: Record<string, unknown>
   webpOptions: Record<string, unknown>
+  avifOptions: Record<string, unknown>
   blurredOptions: { width?: number; toFormat?: ImageFormat }
 }
 export type FileNode = Node & {
@@ -103,12 +104,14 @@ export async function generateImageData({
   cache,
 }: IImageDataArgs): Promise<IGatsbyImageData | undefined> {
   const {
-    layout = `fixed`,
+    layout = `constrained`,
     placeholder = `blurred`,
     tracedSVGOptions = {},
     transformOptions = {},
     quality,
   } = args
+
+  args.formats = args.formats || [`auto`, `webp`]
 
   const {
     fit = `cover`,
@@ -116,6 +119,23 @@ export async function generateImageData({
   } = transformOptions
 
   const metadata = await getImageMetadata(file, placeholder === `dominantColor`)
+
+  if (layout === `fixed` && !args.width && !args.height) {
+    args.width = metadata.width
+  }
+
+  if (
+    layout !== `fixed` &&
+    !args.maxWidth &&
+    !args.maxHeight &&
+    metadata.width
+  ) {
+    if (layout === `constrained`) {
+      args.maxWidth = metadata.width
+    } else if (layout === `fluid`) {
+      args.maxWidth = Math.round(metadata.width / 2)
+    }
+  }
 
   const formats = new Set(args.formats)
   let useAuto = formats.has(``) || formats.has(`auto`) || formats.size === 0
@@ -128,19 +148,29 @@ export async function generateImageData({
   }
 
   let primaryFormat: ImageFormat | undefined
-  let options: Record<string, unknown> | undefined
   if (useAuto) {
     primaryFormat = normalizeFormat(metadata.format || file.extension)
   } else if (formats.has(`png`)) {
     primaryFormat = `png`
-    options = args.pngOptions
   } else if (formats.has(`jpg`)) {
     primaryFormat = `jpg`
-    options = args.jpgOptions
   } else if (formats.has(`webp`)) {
     primaryFormat = `webp`
-    options = args.webpOptions
+  } else if (formats.has(`avif`)) {
+    reporter.warn(
+      `Image ${file.relativePath} specified only AVIF format, with no fallback format. This will mean your site cannot be viewed in all browsers.`
+    )
+    primaryFormat = `avif`
   }
+
+  const optionsMap = {
+    jpg: args.jpgOptions,
+    png: args.pngOptions,
+    webp: args.webpOptions,
+    avif: args.avifOptions,
+  }
+
+  const options = primaryFormat ? optionsMap[primaryFormat] : undefined
 
   const imageSizes: {
     sizes: Array<number>
@@ -170,7 +200,9 @@ export async function generateImageData({
       toFormat: primaryFormat,
     })
 
-    if (pathPrefix) transform.pathPrefix = pathPrefix
+    if (pathPrefix) {
+      transform.pathPrefix = pathPrefix
+    }
     return transform
   })
 
@@ -213,18 +245,54 @@ export async function generateImageData({
     },
   }
 
+  if (primaryFormat !== `avif` && formats.has(`avif`)) {
+    const transforms = imageSizes.sizes.map(outputWidth => {
+      const width = Math.round(outputWidth)
+      const transform = createTransformObject({
+        quality,
+        ...transformOptions,
+        fit,
+        cropFocus,
+        ...args.avifOptions,
+        width,
+        height: Math.round(width / imageSizes.aspectRatio),
+        toFormat: `avif`,
+      })
+      if (pathPrefix) {
+        transform.pathPrefix = pathPrefix
+      }
+      return transform
+    })
+
+    const avifImages = batchQueueImageResizing({
+      file,
+      transforms,
+      reporter,
+    })
+
+    imageProps.images.sources?.push({
+      srcSet: getSrcSet(avifImages),
+      type: `image/avif`,
+      sizes,
+    })
+  }
+
   if (primaryFormat !== `webp` && formats.has(`webp`)) {
     const transforms = imageSizes.sizes.map(outputWidth => {
       const width = Math.round(outputWidth)
       const transform = createTransformObject({
         quality,
-        ...args.transformOptions,
+        ...transformOptions,
+        fit,
+        cropFocus,
         ...args.webpOptions,
-        tracedSVGOptions,
         width,
         height: Math.round(width / imageSizes.aspectRatio),
         toFormat: `webp`,
       })
+      if (pathPrefix) {
+        transform.pathPrefix = pathPrefix
+      }
       return transform
     })
 
@@ -233,10 +301,9 @@ export async function generateImageData({
       transforms,
       reporter,
     })
-    const webpSrcSet = getSrcSet(webpImages)
 
     imageProps.images.sources?.push({
-      srcSet: webpSrcSet,
+      srcSet: getSrcSet(webpImages),
       type: `image/webp`,
       sizes,
     })
@@ -249,7 +316,9 @@ export async function generateImageData({
       file,
       args: {
         ...options,
-        ...args.transformOptions,
+        ...transformOptions,
+        fit,
+        cropFocus,
         toFormatBase64: args.blurredOptions?.toFormat,
         width: placeholderWidth,
         height: Math.round(placeholderWidth / imageSizes.aspectRatio),