+++ packages/gatsby/src/query/file-parser.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-expressions */
 // @flow
 const fs = require(`fs-extra`)
 const crypto = require(`crypto`)
@@ -42,7 +43,7 @@ const generateQueryName = ({ def, hash, file }) => {
 // taken from `babel-plugin-remove-graphql-queries`, in the future import from
 // there
 function followVariableDeclarations(binding) {
-  const node = binding.path?.node
+  const node = binding?.path?.node
   if (
     node?.type === `VariableDeclarator` &&
     node?.id.type === `Identifier` &&
@@ -91,10 +92,7 @@ async function parseToAst(filePath, fileStr, { parentSpan, addError } = {}) {
         ast = tmp
         break
       } catch (error) {
-        boundActionCreators.queryExtractionGraphQLError({
-          componentPath: filePath,
-        })
-        continue
+        // We emit the actual error below if every transpiled variant fails to parse
       }
     }
     if (ast === undefined) {
@@ -376,7 +374,7 @@ async function findGraphQLTags(
                 const binding = followVariableDeclarations(
                   path.scope.getBinding(path.node.local.name)
                 )
-                binding.path.traverse({ TaggedTemplateExpression })
+                binding?.path?.traverse({ TaggedTemplateExpression })
               },
             })
           },